Session Added on every page

+++ delete.php
@@ -1,7 +1,13 @@
 <?php
-include("init.php");
+include "init.php";
+include "validation.php";
 
-$query = $source->Query("DELETE FROM user where id=?",[$_GET['deleteuser']]);
+if (!isset($_SESSION['username'])) {
+  header("Location: login.php");
+}
+else{
+
+    $query = $source->Query("DELETE FROM user where id=?",[$_GET['deleteuser']]);
 if($query){
     $_SESSION['delete_user'] = "";
     header("location:list.php");
@@ -10,5 +16,8 @@ if($query){
     $_SESSION['delete_user'] = "Failed To Delete";
     header("location:list.php");
 }
+}
+
+
 
 ?>
